+++ Data/SBV/Control/Utils.hs
@@ -927,7 +927,10 @@ getAllSatResult = do queryDebug ["*** Checking Satisfiability, all solutions.."]
                      allUninterpreteds <- getUIs
 
                       -- Functions have at least two kinds in their type and all components must be "interpreted"
-                     let allUiFuns = [u | u@(nm, SBVType as) <- allUninterpreteds, length as > 1, not (isNonModelVar cfg nm)]
+                     let allUiFuns = [u | allSatTrackUFs cfg                                      -- config says consider UIFs
+                                        , u@(nm, SBVType as) <- allUninterpreteds, length as > 1  -- get the function ones
+                                        , not (isNonModelVar cfg nm)                              -- make sure they aren't explicitly ignored
+                                     ]
 
                          -- We can only "allSat" if all component types themselves are interpreted. (Otherwise
                          -- there is no way to reflect back the values to the solver.)