+++ Documentation/SBV/Examples/Puzzles/Fish.hs
@@ -74,19 +74,15 @@ mkSymbolicEnumeration ''Sport
 --
 -- It's not hard to modify this program to grab the values of all the assignments, i.e., the full
 -- solution to the puzzle. We leave that as an exercise to the interested reader!
+-- NB. We use the 'allSatTrackUIFuns' configuration to indicate that the uninterpreted function
+-- changes do not matter for generating different values. All we care is that the fishOwner changes!
 fishOwner :: IO ()
-fishOwner = do vs <- getModelValues "fishOwner" `fmap` allSatWith z3{isNonModelVar = modelIgnore} puzzle
+fishOwner = do vs <- getModelValues "fishOwner" `fmap` allSatWith z3{allSatTrackUFs = False} puzzle
                case vs of
                  [Just (v::Nationality)] -> print v
                  []                      -> error "no solution"
                  _                       -> error "no unique solution"
- where -- For model construction purposes, we should not consider
-       -- The different values of the uninterpreted functions we use. We only
-       -- care for one instantiation of that. (Otherwise this'll produce thousands
-       -- of correct but uninteresting results.)
-       modelIgnore s = s `elem` ["color", "nationality", "beverage", "pet", "sport"]
-
-       puzzle = do
+ where puzzle = do
 
           let c = uninterpret "color"
               n = uninterpret "nationality"