+++ CHANGES.md
@@ -49,6 +49,11 @@
   * Added configuration option `allSatPrintAlong`. If set to True, calls to
     allSat will print their models as they are found. The default is False.
 
+  * Added configuration parameter `allSatTrackUFs` (defaulting to True) to control
+    if changes to uninterpreted functions is considered as a different model. In
+    theory this should always be True, but for most practical problems we typically
+    don't care about the function itself but that it exists.
+
   * Added unnamed equivalents of 'sBool', 'sWord8' etc; with a following underscore, i.e.,
     'sBool_', 'sWord8_'. The new functions are supported for all base types, chars,
     strings, lists, and tuples.